Move DatabaseMapper to ultratenant.multidb
