+++ ultratenant/multidb.py
@@ -1,3 +1,5 @@
+from dj_database_url import parse as dburl
+from django.apps import apps
 from django.db import models
 
 from ultratenant.threadlocal import TENANTLOCAL
@@ -12,6 +14,27 @@ class TenantAbstract(models.Model):
 
     def __str__(self):
         return self.key
+
+
+class DatabaseMapper(dict):
+    def __init__(self, default, **others):
+        self.cache = {"default": dburl(default)}
+        self.cache.update({k: dburl(v) for k, v in others.items()})
+
+    def __getitem__(self, key):
+        if key not in self.cache:
+            self.cache[key] = dburl(self.load_dburl(key))
+
+        return self.cache[key]
+
+    def __contains__(self, key):
+        return True
+
+    @staticmethod
+    def load_dburl(key):
+        Tenant = apps.get_model("multidb", "Tenant")
+        t = Tenant.objects.get(key=key)
+        return t.dburl
 
 
 class SQLiteMapper(dict):
